+++ src/bin/socket-lib.ts
@@ -12,7 +12,7 @@
 
 import process from 'node:process'
 
-import { getDefaultLogger } from '../logger/node'
+import { getDefaultLogger } from '../logger/default'
 
 import { runCheck } from './check'