+++ src/releases/github-downloads.ts
@@ -6,7 +6,7 @@ import process from 'node:process'
 
 import { safeMkdir } from '../fs/safe'
 import { httpDownload } from '../http-request/download'
-import { getDefaultLogger } from '../logger/node'
+import { getDefaultLogger } from '../logger/default'
 import { ErrorCtor } from '../primordials/error'
 import {
   StringPrototypeEndsWith,